refactor(oauth2): restructure oauth2 provider into modular package
- Rename identityprovider package to oauth2provider for clarity
- Extract OAuth2 business logic from coderd/oauth2.go into focused modules:
  - apps.go: OAuth2 app management (CRUD operations)
  - app_secrets.go: OAuth2 app secrets management
  - metadata.go: OAuth2 server and resource metadata endpoints
  - registration.go: RFC 7591/7592 dynamic client registration
- Update route handlers to delegate to oauth2provider functions
- Preserve all existing API endpoints and Swagger documentation
- Fix compilation issues and update middleware references
- All tests passing with zero regressions

This refactoring improves code organization and maintainability while
preserving complete API compatibility.
