+++ coderd/oauth2provider/authorize.go
@@ -1,4 +1,4 @@
-package identityprovider
+package oauth2provider
 
 import (
 	"database/sql"
