+++ coderd/oauth2provider/middleware.go
@@ -1,4 +1,4 @@
-package identityprovider
+package oauth2provider
 
 import (
 	"net/http"
